chore: add template of how to add new cloud provier and cloud resouce
