+++ app/application/src/main/java/com/alipay/application/service/account/utils/PlatformUtils.java
@@ -102,6 +102,8 @@ public class PlatformUtils {
                     map.put("vpcEndpoint", hwsPrivateCredential.getVpcEndpoint());
                     map.put("obsEndpoint", hwsPrivateCredential.getObsEndpoint());
                     break;
+//                case PlatformType.Enum.My_Cloud_Provider:
+//                    [3] ADD_NEW_CLOUD : To adapt the logic of parsing authentication information, you need to add a class corresponding to the cloud platform authentication information
                 default:
                     throw new IllegalStateException("Unexpected value: " + platform);
             }
