+++ app/common/src/main/java/com/alipay/common/enums/PlatformType.java
@@ -25,7 +25,6 @@ package com.alipay.common.enums;
  */
 public enum PlatformType {
 
-
     UNKNOWN(Enum.UNKNOWN, "未知", "UNKNOWN"),
     HUAWEI_CLOUD(Enum.HUAWEI_CLOUD, "华为云", "Huawei Cloud"),
 
@@ -44,6 +43,8 @@ public enum PlatformType {
     GCP(Enum.GCP, "GCP", "GCP"),
 
     AZURE(Enum.AZURE, "AZURE", "AZURE");
+    // [2] ADD_NEW_CLOUD : Add a new cloud provider.
+    // MyCloudProvider(Enum.MyCloudProvider, "我的云", "My Cloud Provider");
 
     private String platform;
     private String cnName;
@@ -106,5 +107,7 @@ public enum PlatformType {
         public static final String AWS = "AWS";
         public static final String GCP = "GCP";
         public static final String AZURE = "AZURE";
+        // [1] ADD_NEW_CLOUD : Add a new cloud provider enum.
+        // public static final String MyCloudProvider = "My_Cloud_Provider";
     }
 }
