+++ client/src/utils/const.tsx
@@ -25,6 +25,8 @@ import LOW_LEVEL_RISK from '@/assets/images/LOW_LEVEL_RISK.png';
 import MEDIUM_LEVEL_RISK from '@/assets/images/MEDIUM_LEVEL_RISK.png';
 import TENCENT_ACCOUNT from '@/assets/images/TENCENT_ACCOUNT.png';
 import TENCENT_CLOUD from '@/assets/images/TENCENT_CLOUD.png';
+// [4] ADD_NEW_CLOUD : Add cloud provider icon
+// import My_Cloud_Provider from '@/assets/images/My_Cloud_Provider.png';
 import { FormattedMessage } from '@umijs/max';
 import { TimeRangePickerProps } from 'antd';
 import dayjs from 'dayjs';
@@ -96,6 +98,8 @@ export const platformURLMap = {
   AWS_URL: AWS, // AWS
   GCP_URL: GCP, // Google Cloud
   AZURE_URL: AZURE, // AZURE
+  // [5.1] ADD_NEW_CLOUD : Reference this icon address
+  // My_Cloud_Provider_URL: My_Cloud_Provider
 };
 
 export const overviewURLMap = {
@@ -118,6 +122,8 @@ export const accountURLMap = {
   AWS: AWS_ACCOUNT,
   BAIDU_CLOUD: BAIDU_ACCOUNT,
   AZURE: AZURE_ACCOUNT,
+  // [5.2] ADD_NEW_CLOUD : Reference this icon address
+  // My_Cloud_Provider: My_Cloud_Provider_ACCOUNT
 };
 
 export const RangePresets: TimeRangePickerProps['presets'] = [
