+++ collector/main.go
@@ -34,5 +34,7 @@ func main() {
 		aws.GetPlatformConfig(),
 		tencentcloud.GetPlatformConfig(),
 		baidu.GetPlatformConfig(),
+		// [7] ADD_NEW_CLOUD : Add new cloud provider config
+		//template.GetPlatformConfig(),
 	)
 }
