+++ collector/template/go.mod
@@ -1,4 +1,4 @@
-module test
+module template
 
 go 1.23.0
